rework the API to eliminate the `.execute()` method
Before, the query builder was created by a `.find` or `.find_all`
method, methods were called on the builder, then `.execute()` was
called. Now, the necessary query builder methods are included on `Soup`
and `rcdom::Handle` so that you can just start the query directly from
those. Then, when the query is ready, you call `.find` or `.find_all` to
run the query

Closes #3