Remove the `Clone` requirement for `Query` instances
It was only there for the benefit of the `NodeIterator`, but since
we're already using an `Rc` for other data in that iterator (`Handle`),
it makes sense to use an `Rc` for the queries too. This makes it easier
to implement `Query`s because we don't have to make sure that it's all
`Clone`.