Rustdoc: make first example show the import
When I opened this crate's docs, I started reading it from the top,
found the first example, copied it to a file -- and failed to compile it
because the import was missing. Jumping to the list of modules, it was obvious
to me what to import, but I wish that hurdle wasn't there in the first
place.